requirements.txt
APScheduler==3.10.1
bcrypt==4.0.1
better-profanity==0.7.0
blinker==1.6.2
boto3==1.26.155
botocore==1.29.155
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==40.0.2
cycler==0.11.0
Deprecated==1.2.13
dnspython==2.3.0
email-validator==2.0.0.post2
firebase-admin==6.1.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Dropzone==1.6.0
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
flask-talisman==1.0.0
Flask-WTF==1.1.1
fonttools==4.40.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-logging==3.5.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
kiwisolver==1.4.4
limits==3.5.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.5
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
PyMuPDF==1.22.3
pyotp==2.8.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypng==0.20220715.0
Pyrebase4==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.0.0
pytz==2023.3
qrcode==7.4.2
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.5
rsa==4.9
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.13
stripe==5.4.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
opencv-python
